+++ CHANGELOG.md
@@ -8,6 +8,7 @@
  - Fix "indestructible: false" on wands not working
  - Hunters will mostly usually clean up after themselves on death
  - Some fixes to EMP and SuperEMP
+ - Add allowed-wands and blocked-wands custom region flags
 
 ## 4.9.1